requirements.txt
mastodon.py
python-environ
atproto
Pillow
prompt_toolkit 